use many newer dev-deps
Many things were in beta when this was last published. This PR updates
those packages to use final releases and migrates off `@babel/polyfill`
in favor of `core-js` and `regenerator-runtime`.

It also updates `mocha` from v4 to v9 for parity with the FOLIO's
`@folio/stripes` `v7`/`@folio/stripes-webpack` `v3` where we are running
`webpack` `v5` and `mocha` `v9`.